__assigned_ordinal fields are longs now

+++ business-logic/src/main/java/com/netflix/vms/transformer/hollowoutput/Phase.java
@@ -117,5 +117,5 @@ public class Phase implements Cloneable {
     }
 
     @SuppressWarnings("unused")
-    private int __assigned_ordinal = -1;
+    private long __assigned_ordinal = -1;
 }
