+++ business-logic/src/main/java/com/netflix/vms/transformer/hollowoutput/StreamDataDescriptor.java
@@ -71,5 +71,5 @@ public class StreamDataDescriptor implements Cloneable {
     }
 
     @SuppressWarnings("unused")
-    private int __assigned_ordinal = -1;
+    private long __assigned_ordinal = -1;
 }
